+++ .github/workflows/sdk.cuda.yml
@@ -145,13 +145,9 @@ jobs:
 
       - name: Generate sweep JSON files if missing
         if: ${{ inputs.sweep_root || inputs.sweep_internal || inputs.sweep_evm }}
-        working-directory: crates/sdk
         run: |
-          if [[ ! -f tests/internal_sweep_no_root.json || ! -f tests/internal_sweep_root.json ]]; then
-            cargo nextest run --cargo-profile=fast --features cuda -- generate_interesting_internal_params
-          fi
-          if [[ ! -f tests/root_params.json ]]; then
-            cargo nextest run --cargo-profile=fast --features cuda -- generate_interesting_root_params
+          if [[ ! -f tests/internal_sweep_no_root.json || ! -f tests/internal_sweep_root.json || ! -f tests/root_params.json ]]; then
+            cargo run --bin generate-param-sweep -- crates/sdk/tests/
           fi
 
       - name: Run SDK CUDA tests
@@ -182,10 +178,10 @@ jobs:
             run_sweep "sweeps/sweep_internal" "${SWEEP_ARGS[@]}" --sweep tests/internal_sweep_no_root.json
           fi
           if [[ "${{ inputs.sweep_root || false }}" == "true" ]]; then
-            run_sweep "sweeps/sweep_root" --use-root "${SWEEP_ARGS[@]}" --sweep tests/internal_sweep_root.json
+            run_sweep "sweeps/sweep_root" --use-root "${SWEEP_ARGS[@]}" --sweep tests/internal_sweep_root.json tests/root_params.json
           fi
           if [[ "${{ inputs.sweep_evm || false }}" == "true" ]]; then
-            run_sweep "sweeps/sweep_evm" --use-evm --use-root "${SWEEP_ARGS[@]}" --sweep tests/internal_sweep_root.json
+            run_sweep "sweeps/sweep_evm" --use-evm --use-root "${SWEEP_ARGS[@]}" --sweep tests/internal_sweep_root.json tests/root_params.json
           fi
 
           if [ ${#LOGS[@]} -eq 0 ]; then
