+++ crates/sdk/src/prover/agg.rs
@@ -51,15 +51,6 @@ pub struct InternalLayerMetadata {
 }
 
 impl AggProver {
-    fn assert_internal_recursive_vk_compatible(
-        internal_for_leaf_prover: &InnerAggregationProver<MAX_NUM_CHILDREN_INTERNAL>,
-        internal_recursive_prover: &InnerAggregationProver<MAX_NUM_CHILDREN_INTERNAL>,
-    ) {
-        let internal_for_leaf_vk = internal_for_leaf_prover.get_vk();
-        let internal_recursive_vk = internal_recursive_prover.get_vk();
-        assert_recursive_self_vk_compatible(&internal_for_leaf_vk, &internal_recursive_vk);
-    }
-
     pub fn keygen_prefix(
         app_or_def_vk: Arc<MultiStarkVerifyingKey<SC>>,
         agg_config: AggregationConfig,
@@ -122,10 +113,6 @@ impl AggProver {
             true,
             def_hook_cached_commit,
         );
-        Self::assert_internal_recursive_vk_compatible(
-            &internal_for_leaf_prover,
-            &internal_recursive_prover,
-        );
         Self {
             leaf_prover,
             internal_for_leaf_prover,
@@ -165,10 +152,6 @@ impl AggProver {
             agg_pk.internal_recursive,
             true,
             def_hook_cached_commit,
-        );
-        Self::assert_internal_recursive_vk_compatible(
-            &internal_for_leaf_prover,
-            &internal_recursive_prover,
         );
         Self {
             leaf_prover,
